integration with email and refactoring
