--- experiment/README.md
@@ -1,3 +1,0 @@
-# Experiment
-
-Sevtion to try out the ideas
