--- experiment/docs/README.md
@@ -1,3 +1,0 @@
-# Research 
-
-Ideas and options to tackle the problem
