--- product/README.md
@@ -1,3 +1,0 @@
-# Product 
-
-Here is our final product
